requirements.txt
attrs==23.2.0
cattrs
nodeenv==1.8.0
pytest==8.1.1
pytest-base-url==2.1.0
requests~=2.32.3
virtualenv==20.25.1
Faker~=26.0.0
python-dotenv~=1.0.1
testit-adapter-pytest
testit-api-client
testit-python-commons